Move out ImmutableQueue types
