+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -66,6 +66,8 @@
     <Compile Include="System\Collections\Immutable\ImmutableList_1.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableQueue.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableQueue_1.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableQueue_1.DebuggerProxy.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableQueue_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableSortedDictionary.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableSortedDictionary_2.Builder.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableSortedDictionary_2.cs" />
