+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableQueue_1.DebuggerProxy.cs
@@ -1,0 +1,51 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+using System.Linq;
+
+namespace System.Collections.Immutable
+{
+    /// <summary>
+    /// A simple view of the immutable collection that the debugger can show to the developer.
+    /// </summary>
+    internal class ImmutableQueueDebuggerProxy<T>
+    {
+        /// <summary>
+        /// The collection to be enumerated.
+        /// </summary>
+        private readonly ImmutableQueue<T> _queue;
+
+        /// <summary>
+        /// The simple view of the collection.
+        /// </summary>
+        private T[] _contents;
+
+        /// <summary>   
+        /// Initializes a new instance of the <see cref="ImmutableQueueDebuggerProxy{T}"/> class.
+        /// </summary>
+        /// <param name="queue">The collection to display in the debugger</param>
+        public ImmutableQueueDebuggerProxy(ImmutableQueue<T> queue)
+        {
+            _queue = queue;
+        }
+
+        /// <summary>
+        /// Gets a simple debugger-viewable collection.
+        /// </summary>
+        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
+        public T[] Contents
+        {
+            get
+            {
+                if (_contents == null)
+                {
+                    _contents = _queue.ToArray();
+                }
+
+                return _contents;
+            }
+        }
+    }
+}
